requirements.txt
python-telegram-bot==13.15
SQLAlchemy==2.0.23
pandas==2.1.1
schedule==1.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openpyxl==3.1.2
psutil==5.9.6

# Для адміністративної панелі
Flask==2.3.3
Flask-Login==0.6.2
WTForms==3.0.1
Flask-WTF==1.1.1

# Для візуалізації
matplotlib==3.8.0

# Для тестування
pytest==7.4.2